#1 - Refactored namespaces and folder structure to match solution layout.
