--- src/Core/Application.Batch.Application/Contracts/Persistence/IAddressRepository.cs
@@ -1,8 +1,0 @@
-﻿using Application.Batch.Domain.Entities;
-
-namespace Application.Batch.Application.Contracts.Persistence;
-
-public interface IAddressRepository : IRepository<Address>
-{
-	
-}
