--- src/Core/Application.Batch.Application/Contracts/Persistence/ICustomerRepository.cs
@@ -1,8 +1,0 @@
-﻿using Application.Batch.Domain.Entities;
-
-namespace Application.Batch.Application.Contracts.Persistence;
-
-public interface ICustomerRepository : IRepository<Customer>
-{
-	Task<bool> IsCustomerSocialSecurityNumberUnique(string ssn);
-}
