--- src/Core/Application.Batch.Application/Exceptions/BadRequestException.cs
@@ -1,3 +1,0 @@
-﻿namespace Application.Batch.Application.Exceptions;
-
-public class BadRequestException(string message) : Exception(message);
