--- src/Core/Application.Batch.Application/Exceptions/NotFoundException.cs
@@ -1,3 +1,0 @@
-﻿namespace Application.Batch.Application.Exceptions;
-
-public class NotFoundException(string name, object key) : Exception($"{name} ({key}) is not found.");
