--- src/Core/Application.Batch.Application/Exceptions/ValidationException.cs
@@ -1,18 +1,0 @@
-﻿using FluentValidation.Results;
-
-namespace Application.Batch.Application.Exceptions;
-
-public class ValidationException : Exception
-{
-	public List<string> ValidationErrors { get; set; }
-
-	public ValidationException(ValidationResult validationResult)
-	{
-		ValidationErrors = [];
-
-		foreach (ValidationFailure? validationError in validationResult.Errors)
-		{
-			ValidationErrors.Add(validationError.ErrorMessage);
-		}
-	}
-}
