--- src/Core/Application.Batch.Application/Features/Customers/Commands/CreateCustomer/CreateCustomerCommand.cs
@@ -1,10 +1,0 @@
-﻿using MediatR;
-
-namespace Application.Batch.Application.Features.Customers.Commands.CreateCustomer;
-
-public class CreateCustomerCommand : IRequest<int>
-{
-	public string FirstName { get; set; } = string.Empty;
-	public string LastName { get; set; } = string.Empty;
-	public string SocialSecurityNumber { get; set; } = string.Empty;
-}
