--- src/Core/Application.Batch.Application/Features/Customers/Commands/RemoveCustomer/RemoveCustomerCommand.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-
-namespace Application.Batch.Application.Features.Customers.Commands.RemoveCustomer;
-
-public class RemoveCustomerCommand : IRequest
-{
-	public int Id { get; set; }
-}
