--- src/Core/Application.Batch.Application/Features/Customers/Queries/GetCustomersList/CustomerListViewModel.cs
@@ -1,8 +1,0 @@
-﻿namespace Application.Batch.Application.Features.Customers.Queries.GetCustomersList;
-
-public class CustomerListViewModel
-{
-	public string FirstName { get; set; } = string.Empty;
-	public string LastName { get; set; } = string.Empty;
-	public string SocialSecurityNumber { get; set; } = string.Empty;
-}
