--- src/Core/Application.Batch.Application/Features/Customers/Queries/GetCustomersList/GetCustomerListQuery.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-
-namespace Application.Batch.Application.Features.Customers.Queries.GetCustomersList
-{
-	public class GetCustomerListQuery : IRequest<List<CustomerListViewModel>>
-	{
-	}
-}
