+++ src/Core/Application.Batch.Core.Application/ApplicationServiceRegistration.cs
