+++ src/Core/Application.Batch.Core.Application/Contracts/Persistence/IAddressRepository.cs
@@ -1,0 +1,8 @@
+﻿using Application.Batch.Core.Domain.Entities;
+
+namespace Application.Batch.Core.Application.Contracts.Persistence;
+
+public interface IAddressRepository : IRepository<Address>
+{
+	
+}
