+++ src/Core/Application.Batch.Core.Application/Exceptions/BadRequestException.cs
@@ -1,0 +1,3 @@
+﻿namespace Application.Batch.Core.Application.Exceptions;
+
+public class BadRequestException(string message) : Exception(message);
