+++ src/Core/Application.Batch.Core.Application/Exceptions/NotFoundException.cs
@@ -1,0 +1,3 @@
+﻿namespace Application.Batch.Core.Application.Exceptions;
+
+public class NotFoundException(string name, object key) : Exception($"{name} ({key}) is not found.");
