+++ src/Core/Application.Batch.Core.Application/Exceptions/ValidationException.cs
@@ -1,0 +1,18 @@
+﻿using FluentValidation.Results;
+
+namespace Application.Batch.Core.Application.Exceptions;
+
+public class ValidationException : Exception
+{
+	public List<string> ValidationErrors { get; set; }
+
+	public ValidationException(ValidationResult validationResult)
+	{
+		ValidationErrors = [];
+
+		foreach (ValidationFailure? validationError in validationResult.Errors)
+		{
+			ValidationErrors.Add(validationError.ErrorMessage);
+		}
+	}
+}
