+++ src/Core/Application.Batch.Core.Application/Features/Customers/Commands/CreateCustomer/CreateCustomerCommand.cs
@@ -1,0 +1,10 @@
+﻿using MediatR;
+
+namespace Application.Batch.Core.Application.Features.Customers.Commands.CreateCustomer;
+
+public class CreateCustomerCommand : IRequest<int>
+{
+	public string FirstName { get; set; } = string.Empty;
+	public string LastName { get; set; } = string.Empty;
+	public string SocialSecurityNumber { get; set; } = string.Empty;
+}
