+++ src/Core/Application.Batch.Core.Application/Features/Customers/Commands/RemoveCustomer/RemoveCustomerCommand.cs
@@ -1,0 +1,8 @@
+﻿using MediatR;
+
+namespace Application.Batch.Core.Application.Features.Customers.Commands.RemoveCustomer;
+
+public class RemoveCustomerCommand : IRequest
+{
+	public int Id { get; set; }
+}
