+++ src/Core/Application.Batch.Core.Application/Features/Customers/Queries/GetCustomersList/CustomerListViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Application.Batch.Core.Application.Features.Customers.Queries.GetCustomersList;
+
+public class CustomerListViewModel
+{
+	public string FirstName { get; set; } = string.Empty;
+	public string LastName { get; set; } = string.Empty;
+	public string SocialSecurityNumber { get; set; } = string.Empty;
+}
