+++ src/Core/Application.Batch.Core.Application/Features/Customers/Queries/GetCustomersList/GetCustomerListQuery.cs
@@ -1,0 +1,8 @@
+﻿using MediatR;
+
+namespace Application.Batch.Core.Application.Features.Customers.Queries.GetCustomersList
+{
+	public class GetCustomerListQuery : IRequest<List<CustomerListViewModel>>
+	{
+	}
+}
