+++ src/Core/Application.Batch.Core.Domain/Application.Batch.Core.Domain.csproj
