+++ src/Core/Application.Batch.Core.Domain/Common/AuditableEntity.cs
@@ -1,0 +1,9 @@
+﻿namespace Application.Batch.Core.Domain.Common;
+
+public class AuditableEntity
+{
+	public string? CreatedBy { get; set; }
+	public DateTime CreatedDate { get; set; }
+	public string? LastModifiedBy { get; set; }
+	public DateTime? LastModifiedDate { get; set; }
+}
