+++ src/Core/Application.Batch.Core.Domain/Entities/Customer.cs
@@ -1,0 +1,12 @@
+﻿using Application.Batch.Core.Domain.Common;
+
+namespace Application.Batch.Core.Domain.Entities;
+
+public class Customer : AuditableEntity
+{
+	public int Id { get; set; }
+	public string FirstName { get; set; } = string.Empty;
+	public string LastName { get; set; } = string.Empty;
+	public string SocialSecurityNumber { get; set; } = string.Empty;
+	public List<Address> Addresses { get; set; } = new();
+}
