--- src/Core/Application.Batch.Domain/Common/AuditableEntity.cs
@@ -1,9 +1,0 @@
-﻿namespace Application.Batch.Domain.Common;
-
-public class AuditableEntity
-{
-	public string? CreatedBy { get; set; }
-	public DateTime CreatedDate { get; set; }
-	public string? LastModifiedBy { get; set; }
-	public DateTime? LastModifiedDate { get; set; }
-}
