+++ src/Infrastructure/Application.Batch.Infrastructure.Persistence/Application.Batch.Infrastructure.Persistence.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Core\Application.Batch.Core.Application\Application.Batch.Core.Application.csproj" />
+  </ItemGroup>
+
+</Project>
